+++ app/beds/bed_routes.py
@@ -1,18 +1,17 @@
+from typing import Annotated
 from uuid import UUID
 
-from casbin import AsyncEnforcer
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, Security
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.auth.auth import get_current_active_user
+from app.auth.auth import get_current_user
+from app.auth.auth_schemas import TokenData
 from app.beds.bed_crud import BedCRUD
 from app.beds.bed_models import Bed
 from app.beds.bed_schemas import BedCreate, BedList, BedRead, BedUpdate
-from app.casbin.casbin_config import get_casbin_enforcer
+from app.core.utils.garden_access import check_garden_access
 from app.gardens.garden_crud import GardenCRUD
-from app.helpers.bed_helpers import bed_check_access
 from app.logging import get_logger, log_handler
-from app.users.user_models import User
 from app.utils.database import get_db
 
 logger = get_logger(__name__)
@@ -23,9 +22,10 @@ bed_router = APIRouter(prefix="/beds")
 @bed_router.post("", response_model=Bed)
 async def create_bed(
     bed: BedCreate,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:up", "ga:up:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
-    enforcer: AsyncEnforcer = Depends(get_casbin_enforcer),
 ) -> Bed:
     """
     Route to create a new bed in casbin.
@@ -33,17 +33,12 @@ async def create_bed(
     :param bed: BedCreate object; beds.bed_schemas.BedCreate
     :param session: SQLAlchemy asyncio AsyncSession
     :param current_user: The User accessing the route
-    :param enforcer: Casbin AsyncEnforcer
     :return: Bed; beds.bed_models.Bed
     """
 
-    await bed_check_access(
-        garden_id=bed.garden_id,
-        current_user=current_user,
-        session=session,
-        enforcer=enforcer,
-        action="create",
-    )
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
 
     new_bed = await BedCRUD.create_bed(
         bed=bed,
@@ -68,11 +63,12 @@ async def create_bed(
 @bed_router.get("/garden/{garden_id}", response_model=list[BedList])
 async def read_beds(
     garden_id: UUID,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:re", "ga:re:own"])
+    ],
     skip: int = 0,
     limit: int = 100,
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
-    enforcer: AsyncEnforcer = Depends(get_casbin_enforcer),
 ) -> list[BedList]:
     """
     Route to get all bed routes associated with a given garden.
@@ -82,21 +78,16 @@ async def read_beds(
     :param limit: limit the number of rows
     :param session: SQLAlchemy asyncio AsyncSession
     :param current_user: The User accessing the route
-    :param enforcer: Casbin AsyncEnforcer
     :return: list[BedList]; beds.bed_schema.BedList
     """
-
-    await bed_check_access(
-        garden_id=garden_id,
-        current_user=current_user,
-        session=session,
-        enforcer=enforcer,
-        action="read",
-    )
 
     garden = await GardenCRUD.get_garden(session=session, garden_id=garden_id)
     if not garden:
         raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_garden_access(
+        current_user=current_user, garden_user=garden.user_id, scope="ga:re"
+    )
 
     beds = await BedCRUD.get_beds(
         garden_id=garden_id,
@@ -110,18 +101,17 @@ async def read_beds(
 @bed_router.get("/{bed_id}", response_model=BedRead)
 async def read_bed(
     bed_id: UUID,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:re", "ga:re:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
-    enforcer: AsyncEnforcer = Depends(get_casbin_enforcer),
 ) -> BedRead:
     """
     Route to get a bed route associated with a given garden and bed id.
 
-    :param garden_id: Garden UUID
     :param bed_id: Bed UUID
     :param session: SQLAlchemy asyncio AsyncSession
     :param current_user: User
-    :param enforcer: Casbin AsyncEnforcer
     :return: BedRead; beds.bed_schema.BedRead
     """
 
@@ -129,12 +119,12 @@ async def read_bed(
     if not bed:
         raise HTTPException(status_code=404, detail="Bed not found")
 
-    await bed_check_access(
-        garden_id=bed.garden_id,
-        current_user=current_user,
-        session=session,
-        enforcer=enforcer,
-        action="read",
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_garden_access(
+        current_user=current_user, garden_user=garden.user_id, scope="ga:re"
     )
 
     return bed
@@ -144,9 +134,10 @@ async def read_bed(
 async def update_bed(
     bed_id: UUID,
     bed_update: BedUpdate,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:up", "ga:up:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
-    enforcer: AsyncEnforcer = Depends(get_casbin_enforcer),
 ) -> Bed | None:
     """
     Route to update a bed
@@ -155,7 +146,6 @@ async def update_bed(
     :param bed_update: BedUpdate object; beds.bed_schemas.BedUpdate
     :param session: SQLAlchemy asyncio AsyncSession
     :param current_user: User
-    :param enforcer: Casbin AsyncEnforcer
     :return: Bed; beds.bed_models.Bed
     """
 
@@ -163,12 +153,12 @@ async def update_bed(
     if not bed:
         raise HTTPException(status_code=404, detail="Bed not found")
 
-    await bed_check_access(
-        garden_id=bed.garden_id,
-        current_user=current_user,
-        session=session,
-        enforcer=enforcer,
-        action="update",
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_garden_access(
+        current_user=current_user, garden_user=garden.user_id, scope="ga:up"
     )
 
     updated_bed = await BedCRUD.update_bed(
@@ -195,9 +185,10 @@ async def update_bed(
 @bed_router.delete("/{bed_id}")
 async def delete_bed(
     bed_id: UUID,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:de", "ga:de:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
-    enforcer: AsyncEnforcer = Depends(get_casbin_enforcer),
 ) -> dict:
     """
     Route to delete a bed
@@ -205,7 +196,6 @@ async def delete_bed(
     :param bed_id: Bed UUID
     :param session: SQLAlchemy asyncio AsyncSession
     :param current_user: User
-    :param enforcer: Casbin AsyncEnforcer
     :return: dict
     """
 
@@ -213,12 +203,12 @@ async def delete_bed(
     if not bed:
         raise HTTPException(status_code=404, detail="Bed not found")
 
-    await bed_check_access(
-        garden_id=bed.garden_id,
-        current_user=current_user,
-        session=session,
-        enforcer=enforcer,
-        action="delete",
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_garden_access(
+        current_user=current_user, garden_user=garden.user_id, scope="ga:de"
     )
 
     if not await BedCRUD.delete_bed(session, bed_id):
